codex: extract agent role loader module
Move the new agent-role loader, parser, and discovery logic into a dedicated config module so mod.rs stays focused on config types and entry points.
